% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deps.R
\name{skald_index_status}
\alias{skald_index_status}
\title{Inspect index or store status}
\usage{
skald_index_status(x)
}
\arguments{
\item{x}{A \link{SkaldIndex} or \link{SkaldStore}.}
}
\value{
A tibble describing the index or store.
}
\description{
Reports basic status information for a \link{SkaldIndex} or \link{SkaldStore}.
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
skald_index_status(index)
\dontshow{\}) # examplesIf}
}
