% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deps.R
\name{skald_sitrep}
\alias{skald_sitrep}
\title{Print a skald situation report}
\usage{
skald_sitrep()
}
\value{
Invisibly, a diagnostics list.
}
\description{
Prints a human-readable report for the active R and Python environment and
invisibly returns the same diagnostic data as \code{\link[=skald_diagnostics]{skald_diagnostics()}}.
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
skald_sitrep()
\dontshow{\}) # examplesIf}
}
