% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/store.R
\name{skald_store_remove}
\alias{skald_store_remove}
\title{Mark chunks as removed}
\usage{
skald_store_remove(store, filter)
}
\arguments{
\item{store}{A writable \link{SkaldStore}.}

\item{filter}{Data-masking expression evaluated against stored chunk metadata.}
}
\value{
Invisibly, \code{store}.
}
\description{
Marks chunks matching a filter expression as deleted in a writable
\link{SkaldStore}.
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
skald_store_remove(store, chunk_id == 1)
\dontshow{\}) # examplesIf}
}
